Add drive rovers and collission avoidance.
Also introduce some custom exceptions that is used to detect collisions and then use these exceptions for when falling of acliff as well.
Added tests for most of these. Some of these tests are currently broken but will be fixed asap once the code settles (Changing my mind a lot at this stage).
Some other minor changes are made to methods to improve it. Things like adding a While true loop instead of recalling the method.