Introduce 2 levels for our mission.
Mission now has a beginner and advanced level. Beginner level takes you through the basic case where advanced through all the other work done.
Tests created for this as well, however not for the advanced case. This is due to the iterative nature of it, things get complicated. Not doing this for now.
Also made the deployment a bit smarter to prevent collisions.